+++ src/lib.rs
@@ -136,12 +136,13 @@
 #![cfg_attr(not(test), no_std)]
 
 mod private;
+mod unrestricted;
+mod iterator;
+
+pub use unrestricted::*;
+pub use iterator::*;
 
 use core::marker::PhantomData;
-use core::mem::replace;
-use generic_array::{sequence::GenericSequence, ArrayLength, GenericArray};
-
-use private::Entry;
 
 pub use generic_array::typenum::consts;
 
@@ -157,10 +158,6 @@ pub struct Key<IT, N> {
     _item_marker: PhantomData<IT>,
     _size_marker: PhantomData<N>,
 }
-
-/// Alias of [`ArrayLength`](../generic_array/trait.ArrayLength.html)
-pub trait Size<I>: ArrayLength<Entry<I>> {}
-impl<T, I> Size<I> for T where T: ArrayLength<Entry<I>> {}
 
 impl<IT, N> Key<IT, N> {
     fn new(owner: &Slots<IT, N>, idx: usize) -> Self
@@ -181,15 +178,6 @@ impl<IT, N> Key<IT, N> {
     }
 }
 
-pub struct UnrestrictedSlots<IT, N>
-where
-    N: Size<IT>,
-{
-    items: GenericArray<Entry<IT>, N>,
-    next_free: usize,
-    count: usize,
-}
-
 /// Data type that stores values and returns a key that can be used to manipulate
 /// the stored values.
 /// Values can be read by anyone but can only be modified using the key.
@@ -203,81 +191,6 @@ where
     inner: UnrestrictedSlots<IT, N>,
 }
 
-/// Read-only iterator to access all occupied slots.
-pub struct Iter<'a, IT> {
-    inner: core::slice::Iter<'a, private::Entry<IT>>,
-}
-
-/// Read-write iterator to access all occupied slots.
-pub struct IterMut<'a, IT> {
-    inner: core::slice::IterMut<'a, private::Entry<IT>>,
-}
-
-impl<'a, IT> Iterator for Iter<'a, IT> {
-    type Item = &'a IT;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some(slot) = self.inner.next() {
-            if let Entry::Used(ref item) = slot {
-                return Some(item);
-            }
-        }
-        None
-    }
-}
-
-impl<'a, IT> Iterator for IterMut<'a, IT> {
-    type Item = &'a mut IT;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some(slot) = self.inner.next() {
-            if let Entry::Used(ref mut item) = slot {
-                return Some(item);
-            }
-        }
-        None
-    }
-}
-
-#[cfg(test)]
-mod iter_test {
-    use super::{consts::U3, Slots, UnrestrictedSlots};
-
-    #[test]
-    fn sanity_check() {
-        let mut slots: Slots<_, U3> = Slots::new();
-
-        let _k1 = slots.store(1).unwrap();
-        let k2 = slots.store(2).unwrap();
-        let _k3 = slots.store(3).unwrap();
-
-        slots.take(k2);
-
-        let mut iter = slots.iter();
-        // iterator does not return elements in order of store
-        assert_eq!(Some(&3), iter.next());
-        assert_eq!(Some(&1), iter.next());
-        assert_eq!(None, iter.next());
-
-        for &_ in slots.iter() {}
-    }
-
-    #[test]
-    fn test_mut() {
-        let mut slots: UnrestrictedSlots<_, U3> = UnrestrictedSlots::new();
-
-        let _k1 = slots.store(1).unwrap();
-        let k2 = slots.store(2).unwrap();
-        let _k3 = slots.store(3).unwrap();
-
-        for k in slots.iter_mut() {
-            *k *= 2;
-        }
-
-        assert_eq!(Some(4), slots.take(k2));
-    }
-}
-
 #[cfg(feature = "verify_owner")]
 fn new_instance_id() -> usize {
     use core::sync::atomic::{AtomicUsize, Ordering};
@@ -285,142 +198,6 @@ fn new_instance_id() -> usize {
     static COUNTER: AtomicUsize = AtomicUsize::new(0);
 
     COUNTER.fetch_add(1, Ordering::Relaxed)
-}
-
-impl<IT, N> Default for UnrestrictedSlots<IT, N>
-where
-    N: Size<IT>,
-{
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<IT, N> UnrestrictedSlots<IT, N>
-where
-    N: Size<IT>,
-{
-    /// Creates a new, empty Slots object.
-    pub fn new() -> Self {
-        Self {
-            items: GenericArray::generate(|i| {
-                i.checked_sub(1)
-                    .map(Entry::EmptyNext)
-                    .unwrap_or(Entry::EmptyLast)
-            }),
-            next_free: N::USIZE.saturating_sub(1), // edge case: N == 0
-            count: 0,
-        }
-    }
-
-    /// Returns a read-only iterator.
-    /// The iterator can be used to read data from all occupied slots.
-    ///
-    /// **Note:** Do not rely on the order in which the elements are returned.
-    pub fn iter(&self) -> Iter<IT> {
-        Iter {
-            inner: self.items.iter(),
-        }
-    }
-
-    pub fn iter_mut(&mut self) -> IterMut<IT> {
-        IterMut {
-            inner: self.items.iter_mut(),
-        }
-    }
-
-    /// Returns the number of slots
-    pub fn capacity(&self) -> usize {
-        N::USIZE
-    }
-
-    /// Returns the number of occupied slots
-    pub fn count(&self) -> usize {
-        self.count
-    }
-
-    fn full(&self) -> bool {
-        self.count == self.capacity()
-    }
-
-    fn free(&mut self, idx: usize) {
-        debug_assert!(self.count != 0, "Free called on an empty collection");
-
-        self.items[idx] = if self.full() {
-            Entry::EmptyLast
-        } else {
-            Entry::EmptyNext(self.next_free)
-        };
-
-        self.next_free = idx; // the freed element will always be the top of the free stack
-        self.count -= 1;
-    }
-
-    fn alloc(&mut self) -> Option<usize> {
-        if self.full() {
-            // no free slot
-            None
-        } else {
-            // next_free points to the top of the free stack
-            let index = self.next_free;
-
-            self.next_free = match self.items[index] {
-                Entry::EmptyNext(n) => n, // pop the stack
-                Entry::EmptyLast => 0,    // replace last element with anything
-                _ => unreachable!("Non-empty item in entry behind free chain"),
-            };
-            self.count += 1;
-            Some(index)
-        }
-    }
-
-    /// Store an element in a free slot and return the key to access it.
-    pub fn store(&mut self, item: IT) -> Result<usize, IT> {
-        match self.alloc() {
-            Some(i) => {
-                self.items[i] = Entry::Used(item);
-                Ok(i)
-            }
-            None => Err(item),
-        }
-    }
-
-    /// Remove and return the element that belongs to the key.
-    pub fn take(&mut self, key: usize) -> Option<IT> {
-        if let Entry::Used(item) = replace(&mut self.items[key], Entry::EmptyLast) {
-            self.free(key);
-            Some(item)
-        } else {
-            None
-        }
-    }
-
-    /// Read the element that belongs to a particular index. Since the index may point to
-    /// a free slot or outside the collection, this operation may return None without invoking the callback.
-    ///
-    /// This operation does not move ownership so the `function` callback must be used
-    /// to access the stored element. The callback may return arbitrary derivative of the element.
-    pub fn read<T>(&self, key: usize, function: impl FnOnce(&IT) -> T) -> Option<T> {
-        if key >= self.capacity() {
-            None
-        } else {
-            match &self.items[key] {
-                Entry::Used(item) => Some(function(&item)),
-                _ => None,
-            }
-        }
-    }
-
-    /// Access the element that belongs to the key for modification.
-    ///
-    /// This operation does not move ownership so the `function` callback must be used
-    /// to access the stored element. The callback may return arbitrary derivative of the element.
-    pub fn modify<T>(&mut self, key: usize, function: impl FnOnce(&mut IT) -> T) -> Option<T> {
-        match self.items[key] {
-            Entry::Used(ref mut item) => Some(function(item)),
-            _ => None,
-        }
-    }
 }
 
 impl<IT, N> Slots<IT, N>
@@ -459,11 +236,7 @@ where
 
     /// Returns the number of occupied slots
     pub fn count(&self) -> usize {
-        self.inner.count
-    }
-
-    fn full(&self) -> bool {
-        self.inner.full()
+        self.inner.count()
     }
 
     /// Store an element in a free slot and return the key to access it.
